// login-from1/usingts/src/interfaces.ts
export interface HttpRequestOptions {
    method: string;
    endpoint: string;
    headers: {[key: string]: string}; //metadata ???
    requestBody?: string;
    responseType: XMLHttpRequestResponseType;
}

export interface LoginFormOptions {
    url: string;
    success: () => void;
    error: () => void;
}

export interface LoginApiResponse {
    statusCode: number;
    message: string;
}